fix/#68 : lambda response 설정
